tree: refactor tree.c, filters.c, and lib.c
Move existing linux specific code into separate files, tree-linux.c, filters-linux.c, and lib-linux.c. Create new files with windows stub function for tree-win.c, filters-win.c and lib-win.c. The windows functionality will be implemented in an upcoming change.
